# Use `math/rand` instead of `crypto/rand` in tests to speed up execution 